fix(api): money columns to BigInteger to support amounts > 21M Kz
CRITICAL: int32 INTEGER columns were rejecting any amount above
2,147,483,647 centavos (~21.5M Kz). A real production user tried to
create a 500M Kz debt and got HTTP 500 with asyncpg DataError 'value
out of int32 range'. In Angola, debts above 21M Kz are completely
normal (auto loans, mortgages, etc.).

Migration ALTER TABLE ... TYPE BIGINT applied to 50 monetary columns
across 18 tables: accounts, bills, budgets, debts, expense_splits,
finance_settings, finance_snapshots, goals, income_sources, investments,
invoice_lines, invoices, module_addons, payments, plans, promotions,
recurring_rules, subscriptions, transactions.

bigint range = 9.2 quintillion centavos = effectively unlimited for
financial data.

Also includes:
- Alembic baseline applied to staging and production via stamp
  e996d56b33b3 (the previous head). Production was previously created
  via Base.metadata.create_all without alembic_version table.
- Removes Base.metadata.create_all from app startup. Schema is now
  managed exclusively via Alembic migrations per the production safety
  rules. Only the pgvector extension creation remains in startup.
- Note: production CORS env var ALLOWED_ORIGINS updated separately via
  Railway CLI to remove http://localhost:3000.

+++ apps/api/app/models/account.py
@@ -1,6 +1,6 @@
 import uuid
 
-from sqlalchemy import Boolean, ForeignKey, Integer, SmallInteger, String, Text
+from sqlalchemy import BigInteger, Boolean, ForeignKey, Integer, SmallInteger, String, Text
 from sqlalchemy.dialects.postgresql import ENUM
 from sqlalchemy.dialects.postgresql import UUID as PG_UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
@@ -28,15 +28,15 @@ class Account(BaseModel):
         default=CurrencyCode.AOA,
     )
     # Balance in centavos (integer) to avoid float precision issues
-    balance: Mapped[int] = mapped_column(Integer, default=0)
+    balance: Mapped[int] = mapped_column(BigInteger, default=0)
     icon: Mapped[str | None] = mapped_column(String(10))
     color: Mapped[str | None] = mapped_column(String(7))
     institution: Mapped[str | None] = mapped_column(String(100))
     account_number: Mapped[str | None] = mapped_column(Text)
     # Limite de crédito em centavos (para cartões de crédito)
-    credit_limit: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    credit_limit: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     # Alerta de saldo baixo em centavos
-    low_balance_alert: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    low_balance_alert: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     usage_type: Mapped[AccountUsageType | None] = mapped_column(
         ENUM(AccountUsageType, name="account_usage_type", create_type=True), nullable=True
     )
